feat: collapse whitespace in multiline sql (#20)
When used in conjunction with the colorizing output, only the first line of the SQL is colored. This is due to the logging filter assuming anything after the first line is a traceback; something to present in without color. Unfortunately, this breaks the intent of the SQL logging.

Linebreaks in SQL logging aren't important. Getting rid of them doesn't hurt readability. In fact, even without the colorization, readability is improved because more lines fit on one screen.

This does not catch all SQL with line breaks... I'll have to update the logic again later if I care.